requirements.txt
streamlit
langchain
python-dotenv
PyPDF2